requirements.txt
beautifulsoup4==4.12.3
fake-useragent==1.5.1
PyYAML==6.0.1
requests==2.31.0
requests_toolbelt==1.0.0
schedule==1.2.1
urllib3==1.26.9
itchat-uos==1.5.0.dev0
